LayoutTests/http/tests/cookies/same-site/popup-cross-site-post-expected.txt


PASS '127.0.0.1' is not same-site with 'localhost', so samesite cookies are not sent via POST.
